requirements.txt
# Runtime dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
pydantic>=2.4.2
httpx>=0.25.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
starlette>=0.27.0
bcrypt>=4.0.1
python-jose>=3.3.0
cryptography>=41.0.0
requests>=2.31.0

# Development dependencies
colorama>=0.4.6
pytest>=7.0.0
pytest-asyncio>=0.25.0  # Required for async tests
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
